third_party/blink/web_tests/external/wpt/css/css-color/parsing/color-valid-lab-expected.txt

This is a testharness.js-based test.
Found 92 tests; 12 PASS, 80 FAIL, 0 TIMEOUT, 0 NOTRUN.
FAIL e.style['color'] = "lab(0 0 0)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL e.style['color'] = "lab(0 0 0 / 1)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL e.style['color'] = "lab(0 0 0 / 0.5)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0 / 0.5)" but got "lab(0% 0 0 / 0.5)"
FAIL e.style['color'] = "lab(20 0 10/0.5)" should set the property value assert_equals: serialization should be canonical expected "lab(20 0 10 / 0.5)" but got "lab(20% 0 10 / 0.5)"
FAIL e.style['color'] = "lab(20 0 10/50%)" should set the property value assert_equals: serialization should be canonical expected "lab(20 0 10 / 0.5)" but got "lab(20% 0 10 / 0.5)"
FAIL e.style['color'] = "lab(400 0 10/50%)" should set the property value assert_equals: serialization should be canonical expected "lab(400 0 10 / 0.5)" but got "lab(400% 0 10 / 0.5)"
FAIL e.style['color'] = "lab(50 -160 160)" should set the property value assert_equals: serialization should be canonical expected "lab(50 -160 160)" but got "lab(50% -160 160)"
FAIL e.style['color'] = "lab(50 -200 200)" should set the property value assert_equals: serialization should be canonical expected "lab(50 -200 200)" but got "lab(50% -200 200)"
FAIL e.style['color'] = "lab(0 0 0 / -10%)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0 / 0)" but got "lab(0% 0 0 / 0)"
FAIL e.style['color'] = "lab(0 0 0 / 110%)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL e.style['color'] = "lab(0 0 0 / 300%)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL e.style['color'] = "lab(-40 0 0)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0)" but got "lab(0% 0 0)"
FAIL e.style['color'] = "lab(50 -20 0)" should set the property value assert_equals: serialization should be canonical expected "lab(50 -20 0)" but got "lab(50% -20 0)"
FAIL e.style['color'] = "lab(50 0 -20)" should set the property value assert_equals: serialization should be canonical expected "lab(50 0 -20)" but got "lab(50% 0 -20)"
FAIL e.style['color'] = "lab(calc(50 * 3) calc(0.5 - 1) calc(1.5) / calc(-0.5 + 1))" should set the property value assert_equals: serialization should be canonical expected "lab(150 -0.5 1.5 / 0.5)" but got "lab(150% -0.5 1.5 / 0.5)"
FAIL e.style['color'] = "lab(calc(-50 * 3) calc(0.5 + 1) calc(-1.5) / calc(-0.5 * 2))" should set the property value assert_equals: serialization should be canonical expected "lab(0 1.5 -1.5 / 0)" but got "lab(0% 1.5 -1.5 / 0)"
PASS e.style['color'] = "lab(none none none / none)" should set the property value
PASS e.style['color'] = "lab(none none none)" should set the property value
FAIL e.style['color'] = "lab(20 none none / none)" should set the property value assert_equals: serialization should be canonical expected "lab(20 none none / none)" but got "lab(20% none none / none)"
PASS e.style['color'] = "lab(none none none / 0.5)" should set the property value
FAIL e.style['color'] = "lab(0 0 0 / none)" should set the property value assert_equals: serialization should be canonical expected "lab(0 0 0 / none)" but got "lab(0% 0 0 / none)"
FAIL e.style['color'] = "oklab(0 0 0)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL e.style['color'] = "oklab(0 0 0 / 1)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL e.style['color'] = "oklab(0 0 0 / 0.5)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0 / 0.5)" but got "oklab(0% 0 0 / 0.5)"
FAIL e.style['color'] = "oklab(20 0 10/0.5)" should set the property value assert_equals: serialization should be canonical expected "oklab(20 0 10 / 0.5)" but got "oklab(20% 0 10 / 0.5)"
FAIL e.style['color'] = "oklab(20 0 10/50%)" should set the property value assert_equals: serialization should be canonical expected "oklab(20 0 10 / 0.5)" but got "oklab(20% 0 10 / 0.5)"
FAIL e.style['color'] = "oklab(400 0 10/50%)" should set the property value assert_equals: serialization should be canonical expected "oklab(400 0 10 / 0.5)" but got "oklab(400% 0 10 / 0.5)"
FAIL e.style['color'] = "oklab(50 -160 160)" should set the property value assert_equals: serialization should be canonical expected "oklab(50 -160 160)" but got "oklab(50% -160 160)"
FAIL e.style['color'] = "oklab(50 -200 200)" should set the property value assert_equals: serialization should be canonical expected "oklab(50 -200 200)" but got "oklab(50% -200 200)"
FAIL e.style['color'] = "oklab(0 0 0 / -10%)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0 / 0)" but got "oklab(0% 0 0 / 0)"
FAIL e.style['color'] = "oklab(0 0 0 / 110%)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL e.style['color'] = "oklab(0 0 0 / 300%)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL e.style['color'] = "oklab(-40 0 0)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0)" but got "oklab(0% 0 0)"
FAIL e.style['color'] = "oklab(50 -20 0)" should set the property value assert_equals: serialization should be canonical expected "oklab(50 -20 0)" but got "oklab(50% -20 0)"
FAIL e.style['color'] = "oklab(50 0 -20)" should set the property value assert_equals: serialization should be canonical expected "oklab(50 0 -20)" but got "oklab(50% 0 -20)"
FAIL e.style['color'] = "oklab(calc(50 * 3) calc(0.5 - 1) calc(1.5) / calc(-0.5 + 1))" should set the property value assert_equals: serialization should be canonical expected "oklab(150 -0.5 1.5 / 0.5)" but got "oklab(150% -0.5 1.5 / 0.5)"
FAIL e.style['color'] = "oklab(calc(-50 * 3) calc(0.5 + 1) calc(-1.5) / calc(-0.5 * 2))" should set the property value assert_equals: serialization should be canonical expected "oklab(0 1.5 -1.5 / 0)" but got "oklab(0% 1.5 -1.5 / 0)"
PASS e.style['color'] = "oklab(none none none / none)" should set the property value
PASS e.style['color'] = "oklab(none none none)" should set the property value
FAIL e.style['color'] = "oklab(20 none none / none)" should set the property value assert_equals: serialization should be canonical expected "oklab(20 none none / none)" but got "oklab(20% none none / none)"
PASS e.style['color'] = "oklab(none none none / 0.5)" should set the property value
FAIL e.style['color'] = "oklab(0 0 0 / none)" should set the property value assert_equals: serialization should be canonical expected "oklab(0 0 0 / none)" but got "oklab(0% 0 0 / none)"
FAIL e.style['color'] = "lch(0 0 0deg)" should set the property value assert_equals: serialization should be canonical expected "lch(0 0 0)" but got "lch(0% 0 0)"
FAIL e.style['color'] = "lch(0 0 0deg / 1)" should set the property value assert_equals: serialization should be canonical expected "lch(0 0 0)" but got "lch(0% 0 0)"
FAIL e.style['color'] = "lch(0 0 0deg / 0.5)" should set the property value assert_equals: serialization should be canonical expected "lch(0 0 0 / 0.5)" but got "lch(0% 0 0 / 0.5)"
FAIL e.style['color'] = "lch(100 230 0deg / 0.5)" should set the property value assert_equals: serialization should be canonical expected "lch(100 230 0 / 0.5)" but got "lch(100% 230 0 / 0.5)"
FAIL e.style['color'] = "lch(20 50 20deg/0.5)" should set the property value assert_equals: serialization should be canonical expected "lch(20 50 20 / 0.5)" but got "lch(20% 50 20 / 0.5)"
FAIL e.style['color'] = "lch(20 50 20deg/50%)" should set the property value assert_equals: serialization should be canonical expected "lch(20 50 20 / 0.5)" but got "lch(20% 50 20 / 0.5)"
FAIL e.style['color'] = "lch(10 20 20deg / -10%)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20 / 0)" but got "lch(10% 20 20 / 0)"
FAIL e.style['color'] = "lch(10 20 20deg / 110%)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(10 20 1.28rad)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 73.3386)" but got "lch(10% 20 73.3386)"
FAIL e.style['color'] = "lch(10 20 380deg)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(10 20 -340deg)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(10 20 740deg)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(10 20 -700deg)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(-40 0 0)" should set the property value assert_equals: serialization should be canonical expected "lch(0 0 0)" but got "lch(0% 0 0)"
FAIL e.style['color'] = "lch(20 -20 0)" should set the property value assert_equals: serialization should be canonical expected "lch(20 0 0)" but got "lch(20% 0 0)"
FAIL e.style['color'] = "lch(0 0 0 / 0.5)" should set the property value assert_equals: serialization should be canonical expected "lch(0 0 0 / 0.5)" but got "lch(0% 0 0 / 0.5)"
FAIL e.style['color'] = "lch(10 20 20 / 110%)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(10 20 -700)" should set the property value assert_equals: serialization should be canonical expected "lch(10 20 20)" but got "lch(10% 20 20)"
FAIL e.style['color'] = "lch(calc(50 * 3) calc(0.5 - 1) calc(20deg * 2) / calc(-0.5 + 1))" should set the property value assert_equals: serialization should be canonical expected "lch(150 0 40 / 0.5)" but got "lch(150% 0 40 / 0.5)"
FAIL e.style['color'] = "lch(calc(-50 * 3) calc(0.5 + 1) calc(-20deg * 2) / calc(-0.5 * 2))" should set the property value assert_equals: serialization should be canonical expected "lch(0 1.5 320 / 0)" but got "lch(0% 1.5 320 / 0)"
PASS e.style['color'] = "lch(none none none / none)" should set the property value
PASS e.style['color'] = "lch(none none none)" should set the property value
FAIL e.style['color'] = "lch(20 none none / none)" should set the property value assert_equals: serialization should be canonical expected "lch(20 none none / none)" but got "lch(20% none none / none)"
PASS e.style['color'] = "lch(none none none / 0.5)" should set the property value
FAIL e.style['color'] = "lch(0 0 0 / none)" should set the property value assert_equals: serialization should be canonical expected "lch(0 0 0 / none)" but got "lch(0% 0 0 / none)"
FAIL e.style['color'] = "oklch(0 0 0deg)" should set the property value assert_equals: serialization should be canonical expected "oklch(0 0 0)" but got "oklch(0% 0 0)"
FAIL e.style['color'] = "oklch(0 0 0deg / 1)" should set the property value assert_equals: serialization should be canonical expected "oklch(0 0 0)" but got "oklch(0% 0 0)"
FAIL e.style['color'] = "oklch(0 0 0deg / 0.5)" should set the property value assert_equals: serialization should be canonical expected "oklch(0 0 0 / 0.5)" but got "oklch(0% 0 0 / 0.5)"
FAIL e.style['color'] = "oklch(100 230 0deg / 0.5)" should set the property value assert_equals: serialization should be canonical expected "oklch(100 230 0 / 0.5)" but got "oklch(100% 230 0 / 0.5)"
FAIL e.style['color'] = "oklch(20 50 20deg/0.5)" should set the property value assert_equals: serialization should be canonical expected "oklch(20 50 20 / 0.5)" but got "oklch(20% 50 20 / 0.5)"
FAIL e.style['color'] = "oklch(20 50 20deg/50%)" should set the property value assert_equals: serialization should be canonical expected "oklch(20 50 20 / 0.5)" but got "oklch(20% 50 20 / 0.5)"
FAIL e.style['color'] = "oklch(10 20 20deg / -10%)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20 / 0)" but got "oklch(10% 20 20 / 0)"
FAIL e.style['color'] = "oklch(10 20 20deg / 110%)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(10 20 1.28rad)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 73.3386)" but got "oklch(10% 20 73.3386)"
FAIL e.style['color'] = "oklch(10 20 380deg)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(10 20 -340deg)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(10 20 740deg)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(10 20 -700deg)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(-40 0 0)" should set the property value assert_equals: serialization should be canonical expected "oklch(0 0 0)" but got "oklch(0% 0 0)"
FAIL e.style['color'] = "oklch(20 -20 0)" should set the property value assert_equals: serialization should be canonical expected "oklch(20 0 0)" but got "oklch(20% 0 0)"
FAIL e.style['color'] = "oklch(0 0 0 / 0.5)" should set the property value assert_equals: serialization should be canonical expected "oklch(0 0 0 / 0.5)" but got "oklch(0% 0 0 / 0.5)"
FAIL e.style['color'] = "oklch(10 20 20 / 110%)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(10 20 -700)" should set the property value assert_equals: serialization should be canonical expected "oklch(10 20 20)" but got "oklch(10% 20 20)"
FAIL e.style['color'] = "oklch(calc(50 * 3) calc(0.5 - 1) calc(20deg * 2) / calc(-0.5 + 1))" should set the property value assert_equals: serialization should be canonical expected "oklch(150 0 40 / 0.5)" but got "oklch(150% 0 40 / 0.5)"
FAIL e.style['color'] = "oklch(calc(-50 * 3) calc(0.5 + 1) calc(-20deg * 2) / calc(-0.5 * 2))" should set the property value assert_equals: serialization should be canonical expected "oklch(0 1.5 320 / 0)" but got "oklch(0% 1.5 320 / 0)"
PASS e.style['color'] = "oklch(none none none / none)" should set the property value
PASS e.style['color'] = "oklch(none none none)" should set the property value
FAIL e.style['color'] = "oklch(20 none none / none)" should set the property value assert_equals: serialization should be canonical expected "oklch(20 none none / none)" but got "oklch(20% none none / none)"
PASS e.style['color'] = "oklch(none none none / 0.5)" should set the property value
FAIL e.style['color'] = "oklch(0 0 0 / none)" should set the property value assert_equals: serialization should be canonical expected "oklch(0 0 0 / none)" but got "oklch(0% 0 0 / none)"
Harness: the test ran to completion.
